Split out wireSubscriptions from wireMethods, add wirePublications
