+++ imports/api/tasks/methods.js
@@ -6,6 +6,7 @@ const methods = [
   {
     name: 'upsertTask',
     returns: 'Task',
+    // auth: true,
     params: {
       _id: { type: String, optional: true },
       description: { type: String, min: 1 },
@@ -26,6 +27,7 @@ const methods = [
   {
     name: 'setTaskDone',
     returns: 'Task',
+    // auth: true,
     params: {
       _id: { type: String },
       done: { type: Boolean },
@@ -39,6 +41,7 @@ const methods = [
   {
     name: 'removeTask',
     returns: Boolean,
+    // auth: true,
     params: {
       _id: { type: String },
     },
